[groupby] add tests && fix some code lint
