+++ common/datablocks/src/kernels/mod.rs
@@ -5,6 +5,8 @@
 #[cfg(test)]
 mod data_block_concat_test;
 #[cfg(test)]
+mod data_block_group_by_hash_test;
+#[cfg(test)]
 mod data_block_group_by_test;
 #[cfg(test)]
 mod data_block_scatter_test;
@@ -15,9 +17,10 @@ mod data_block_take_test;
 
 mod data_block_concat;
 mod data_block_group_by;
+mod data_block_group_by_hash;
 mod data_block_scatter;
 mod data_block_sort;
 mod data_block_take;
 
-pub use data_block_group_by::*;
+pub use data_block_group_by_hash::*;
 pub use data_block_sort::SortColumnDescription;
